+++ src/modules/product-image/product-image.service.ts
@@ -3,7 +3,12 @@ import { CreateProductImageDto } from './dto/create-product-image.dto';
 import { UpdateProductImageDto } from './dto/update-product-image.dto';
 import { InjectRepository } from '@nestjs/typeorm';
 import { ProductImage } from './entities/product-image.entity';
-import { FindManyOptions, FindOneOptions, Repository } from 'typeorm';
+import {
+  EntityManager,
+  FindManyOptions,
+  FindOneOptions,
+  Repository,
+} from 'typeorm';
 import { FileService } from '../file/file.service';
 import { FileType } from 'src/common/enums/product.enum';
 
@@ -17,7 +22,7 @@ export class ProductImageService {
   async create(dto: CreateProductImageDto) {
     for (const file of dto.files) {
       const image = await this.fileService.save(file, FileType.IMAGE);
-      return await this.productImageRepo.save({
+      await this.productImageRepo.save({
         productId: dto.productId,
         fileId: image.id,
       });
@@ -32,11 +37,15 @@ export class ProductImageService {
     return await this.productImageRepo.findOne(option);
   }
 
-  update(id: string, updateProductImageDto: UpdateProductImageDto) {
-    return `This action updates a #${id} productImage`;
-  }
+  async remove(id: string, manager?: EntityManager) {
+    const productImage = await this.findOne({
+      where: { id },
+      relations: { file: true },
+    });
 
-  async remove(id: string) {
-    return await this.productImageRepo.delete(id);
+    if (!productImage) return;
+
+    await this.fileService.remove(productImage.file.id);
+    await this.productImageRepo.delete(id);
   }
 }
